#Docs : ProductLike 수정

+++ src/main/java/com/project/devgram/service/ProductLikeService.java
@@ -3,13 +3,17 @@ package com.project.devgram.service;
 import com.project.devgram.dto.ProductLikeDto;
 import com.project.devgram.entity.Product;
 import com.project.devgram.entity.ProductLike;
+import com.project.devgram.entity.Users;
+import com.project.devgram.exception.DevGramException;
+import com.project.devgram.exception.errorcode.ProductErrorCode;
+import com.project.devgram.exception.errorcode.UserErrorCode;
 import com.project.devgram.repository.ProductLikeRepository;
 import com.project.devgram.repository.ProductRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
+import com.project.devgram.repository.UserRepository;
 import java.util.List;
 import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
@@ -17,23 +21,41 @@ public class ProductLikeService {
 
 	private final ProductLikeRepository productLikeRepository;
 	private final ProductRepository productRepository;
-	private Product product;
+	private final UserRepository userRepository;
 
 	public void productLike(ProductLikeDto productLikeDto) {
 
+		Product product = productRepository.findById(productLikeDto.getProductSeq()).orElse(null);
+		if (product == null){
+			throw new DevGramException(ProductErrorCode.CANNOT_FIND_PRODUCT_BY_PRODUCT_SEQ);
+		}
+
+		Users users = userRepository.findByUsername(productLikeDto.getUsername()).orElse(null);
+		if (users == null){
+			throw new DevGramException(UserErrorCode.USER_NOT_EXIST);
+		}
+
 		ProductLike productLike = ProductLike.builder()
-			.productSeq(productLikeDto.getProductSeq())
-			.username(productLikeDto.getUsername())
+			.product(product)
+			.users(users)
 			.build();
 		productLikeRepository.save(productLike);
 
-		product = productRepository.findById(productLike.getProductSeq()).get();
 		product.setLikeCount(product.getLikeCount() + 1);
 		productRepository.save(product);
 
 	}
 
 	public void productUnLike(ProductLikeDto productLikeDto) {
+
+		Product product = productRepository.findById(productLikeDto.getProductSeq()).orElse(null);
+		if (product == null){
+			throw new DevGramException(ProductErrorCode.CANNOT_FIND_PRODUCT_BY_PRODUCT_SEQ);
+		}
+		Users users = userRepository.findByUsername(productLikeDto.getUsername()).orElse(null);
+		if (users == null){
+			throw new DevGramException(UserErrorCode.USER_NOT_EXIST);
+		}
 		Optional<ProductLike> productLike = findProductLikeByUsernameAndProductSeq(productLikeDto);
 
 		productLikeRepository.delete(productLike.get());
@@ -45,13 +67,18 @@ public class ProductLikeService {
 
 	public Optional<ProductLike> findProductLikeByUsernameAndProductSeq(
 		ProductLikeDto productLikeDto) {
-		return productLikeRepository.findProductLikeByUsernameAndProductSeq(
+		Product product = productRepository.findById(productLikeDto.getProductSeq()).orElse(null);
+		if (product == null){
+			throw new DevGramException(ProductErrorCode.CANNOT_FIND_PRODUCT_BY_PRODUCT_SEQ);
+		}
+		return productLikeRepository.findProductLikeByUsersUsernameAndProductProductSeq(
 			productLikeDto.getUsername(),
 			productLikeDto.getProductSeq());
 	}
 
 	public List<ProductLikeDto> list(String username) {
-		List<ProductLike> productLikes = productLikeRepository.findAllByUsername(username);
+
+		List<ProductLike> productLikes = productLikeRepository.findAllByUsersUsername(username);
 		return ProductLikeDto.of(productLikes);
 	}
 }
